requirements.txt
absl-py==2.3.1
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.20.1
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
hf-xet==1.2.0
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
iopath==0.1.10
ipykernel==7.1.0
ipython==8.37.0
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mediapipe==0.10.15
ml_dtypes==0.5.4
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
omegaconf==2.3.0
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
patsy==1.0.2
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
portalocker==3.2.0
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.11
pycparser==2.23
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
sam2==1.1.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sounddevice==0.5.3
stack-data==0.6.3
statsmodels==0.14.6
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wcwidth==0.2.14